fix: Remove bigtable instances left over when system tests run (#1265)
* Use the prefix from the common file for deleting

Instances are not currently being cleaned up by the reap instances function because a filter is not selecting instances based on the right prefix. It looks like this fix should solve the issue.

* fix: Use the prefix from the common file for deleting

Instances are not currently being cleaned up by the reap instances function because a filter is not selecting instances based on the right prefix. It looks like this fix should solve the issue.