feat: Add plumbing to support unary calls for client side metric collection (#1631)
## Description

This PR adds source code changes with plumbing so that client side metric collection can be done for unary calls. No client side metrics are actually collected yet for readModifyWriteRow or checkAndMutate calls, but future PRs will use this plumbing to collect metrics for those calls. A test is provided that makes a fake method on the table to demonstrate how a method would collect these metrics to serve as an example.

## Impact

Doesn't change the way the client library works. Just provides support for when we open a PR that collects client side metrics for unary calls.

## Testing

A unit test is created that adds a fake method to a table that does a basic unary readModifyWriteRow call. It ensures that the metrics make it to the test metrics handler correctly.

## Additional Information

**Changes:**

`src/interceptor.ts`: This file contains code for attaching interceptors to a call that will call the correct metrics collector methods when data arrives

`src/client-side-metrics/operation-metrics-collector.ts`: The types are cleaned up in this file. None of the functionality has changed in this file

`system-test/read-modify-write-row-interceptors.ts`: This adds a test that ensures the plumbing for collecting metadata trailers and headers works as intended for unary calls.