feat: Add client side metrics for readModifyWriteRow calls (#1656)
## Description

This PR collects client side metrics for readModifyWriteRow calls and collects all the right types of metrics for these unary calls. They leverage the plumbing provided for unary calls to add interceptors to those calls so that we can record metrics for the right zone/cluster/server time.

## Impact

This PR collects client side metrics for readModifyWriteRow calls.

## Testing

For each readRows/mutateRows/mutateRow test we add a corresponding test for readModifyWriteRow.